Make the static properties into a prototype object for the entity
In Javascript, every object has a prototype pointer, and it will use
the prototype object as a fallback if the property is not found on the
object. This is perfect for defining properties that are shared among
objects, but can be overridden.

LambdaMOO also structured game objects with a prototype chain.