Add render order, and make dead monsters become corpses
Render order is a static property that tells us which entities to draw
first. The problem was that the player would get drawn before a
monster's corpse, and we want the player to always be drawn last.

Dead monsters change their *type* from whatever they were to a
"corpse" type.