Cross-font compatibility
I had previously hard-coded the size of the message area and health
bar to match the size of the canvas, which depends on the size of the
Roboto Mono font on my system. Since not everyone will have Roboto
Mono already installed, and since Safari doesn't allow third party
local fonts, I changed the layout to make the message area and health
bar to match the size of the canvas using the default monospace font.