Change entity location to be a record {x: y:}
This is in preparation for Part 8 of the python tutorial, which covers
items and inventory. Items can be on the ground *or* in the player's
inventory. The Python tutorial handles this by removing them from the
global 'entities' list, and ignoring the x,y fields. I plan to
represent this by not having {x: y:} fields when it's in the player's
inventory, but I will leave it in the entities list, and have the
player's inventory include the ids of the entities that it carries.
Not yet sure this is the wisest move but I'll try it.